==> 2b_Home_Assign RID_Parallelize.ipynb <==
# Home and Work Assign RID

This code is used to re-assign the roadID for each workplace. The distance function in geopandas fialed to generate the correct distance, so I create a new distance function in this code

import geopandas as gpd
import pandas as pd
import numpy as np
import timeit
import multiprocessing
from math import sin, cos, sqrt, atan2, radians

### Load Data

#Home CT
homect = gpd.read_file('../Shp_Study_Area_Richard/Home_Point/Home_By_State/home_ct.shp')
print("CT size", len(homect))
#Home NJ
#Home PA
#homepa = gpd.read_file('../Shp_Study_Area_Richard/Home_Point/Home_By_State/home_pa.shp')
#print("PA size", len(homepa))

homect.head()

#Road
road = gpd.read_file('../Shp_Study_Area_Richard/Road_Point/road_points_Aug_Richard.shp')
road.shape

road['RID'] = road['RDID']

road.head()

## Functions

def new_distance(x1, y1, x2, y2):
    # approximate radius of earth in km
    R = 6373.0
    
    lat1 = radians (x1)
    long1 = radians (y1)
    lat2 = radians(x2)
    long2 = radians(y2)
    
    dlon = long2 - long1
    dlat = lat2 - lat1
    
    a = sin(dlat / 2)**2 + cos(lat1) * cos(lat2) * sin(dlon / 2)**2
    c = 2 * atan2(sqrt(a), sqrt(1 - a))

    distance = R * c
    return distance

def assign_rid(data, road):
    #Geo Location
    d_geometry = data.geometry
    #print(d_geometry)
    
    #Create Buffer for each Home or Work Point
    buff = data.geometry.buffer(0.001)
    # Lat and Long
    x = data.lat
    y = data.long
    #print("Site Lat",x)
    #print("Site Long", y)
    
    #Find Intersected points
    r_in = road[road.intersects(buff)].copy()
    #print("Intersected road points numbers:", len(r_in))

    if r_in.empty:
        #print(data.wrkID, "First Empty")
        buff = data.geometry.buffer(0.02)
        r_in = road[road.intersects(buff)].copy()
        #Intersect Road Point Lat list
        rx = r_in.loc[:,'Lat'].tolist()
        #Intersect Road Point Long list
        ry = r_in.loc[:, 'Long'].tolist()
        
        #Calculate Distance between the point and intersected road points 
        dist = []
        for j in range(0, len(r_in)):
            d = new_distance(x, y, rx[j], ry[j])
            dist.append(d)  
        #print("Distance list's length",len(dist))
            
        rid = r_in.loc[:,'RDID']
        #Create DF
        df_rd = pd.DataFrame({'RDID': rid, 'Dist':dist}).sort_values(by='Dist')
        #print(df_rd.head(1))
        #print(df_rd.iloc[0, 0])
        #r_id.append(df_rd.iloc[0, 0])
        return df_rd.iloc[0, 0]
    
    else:
        #Intersect Road Point Lat list
        rx = r_in.loc[:,'Lat'].tolist()
        #Intersect Road Point Long list
        ry = r_in.loc[:, 'Long'].tolist()
        
        #Calculate Distance between the point and intersected road points 
        dist = []
        for j in range(0, len(r_in)):
            d = new_distance(x, y, rx[j], ry[j])
            dist.append(d)  
        #print("Distance list's length",len(dist))
            
        rid = r_in.loc[:,'RDID']
        #Create DF
        df_rd = pd.DataFrame({'RDID': rid, 'Dist':dist}).sort_values(by='Dist')
        #print(df_rd.head(1))
        #print(df_rd.iloc[0, 0])
        #r_id.append(df_rd.iloc[0, 0])
        return df_rd.iloc[0, 0]

def go(data):
    data['RID'] = data.apply(assign_rid, args=(road,),axis=1)
    return data
    
def parallelize(data, func):
    #Cores
    num_cores = multiprocessing.cpu_count()-1
    num_partitions = num_cores
    #Split
    data_split = np.array_split(data, num_partitions)
    #Pools
    pool = multiprocessing.Pool(num_cores)
    
    df = pd.concat(pool.map(func, data_split))
    return df

    pool.close()
    pool.join()

### Assign Function Test

test = homect.iloc[:10,:].copy()
#test['SrdID'] = np.repeat(0,len(test))

test

test['RID'] = test.apply(assign_rid, args=(road,),axis=1)

test

test.merge(road, on='RID', how='inner')

### Parallelize Test

test = homect.iloc[:50,:].copy()
#test['SrdID'] = np.repeat(0,len(test))

#Test 
start_time = timeit.default_timer()
test_p = parallelize(test, go)
elapsed = timeit.default_timer() - start_time
print(elapsed)

test_p

test_p.merge(road, on='RID', how='inner')

### Apply Full Function

#Apply 
if __name__=='__main__':
    print('Start running...')
    #set timer
    start_time = timeit.default_timer()
    #apply
    #wrk_r = parallelize(wrk, go)
    #pa_r = parallelize(homepa, go)
    ct_r = parallelize(homect, go)
    
    elapsed = timeit.default_timer() - start_time
    print("Total Time(s):", elapsed)
    print('End program')

ct_r.head()

ct_r.iloc[:,[0,-1]].to_csv('home_ct_rid_oct.csv')
