==> 3_Combine_Results.ipynb <==
import pandas as pd
import glob
import os

allFiles = glob.glob(os.path.join ("*.csv"))

len(allFiles)

list_ = []
for file_ in allFiles:
    df = pd.read_csv(file_,index_col=None, header=0)
    list_.append(df)

frame = pd.concat(list_, axis = 0, ignore_index = True)

frame.shape

frame.head()

wrk = frame.drop('Unnamed: 0',1)

wrk.head()

wrk.to_csv('wrk_rid.csv')
