==> 4_Home_find_agent.ipynb <==
import pandas as pd
import geopandas as gpd
import numpy as np
import timeit
import multiprocessing

popPts = pd.read_csv('pop_wp_Aug_Richard.csv').iloc[:,1:]

popPts.head()

#CT Home Point
cthp = pd.read_csv('1_Home_Pts/home_ct_rid.csv').iloc[:,1:]
cthp.head()

#NJ Home Point
njhp = pd.read_csv('1_Home_Pts/home_nj_rid.csv').iloc[:,1:]
njhp.head()

njhp.shape

#Ny Home Point
nyhp = pd.read_csv('1_Home_Pts/home_ny_rid.csv').iloc[:,1:]
nyhp.head()

def match_rdID (hmID, gdf):
    try:
        return (gdf.loc[gdf.hhold == hmID, 'SrdID'].values[0])
    except IndexError:
        pass
      
#def search_gdf (hmID, gdf1, gdf2, gdf3, gdf4, counter):
def search_gdf (hmID, gdf1, counter):
    rdID = 0
    if (rdID == 0):
        rdID = match_rdID(hmID, gdf1)
      #print('insearch ', rdID)
    counter += 1
    return rdID
      
def add_rdIDs(pdf):
    counter = 1
    #start_time = timeit.default_timer()
    #pdf['rdID'] = pdf.apply(lambda row: search_gdf(row['hhold'], hmSplit1, hmSplit2, hmSplit3, hmSplit4, counter), axis=1)
    pdf['RID'] = pdf.apply(lambda row: search_gdf(row['hhold'],nyhp, counter), axis=1)
    #elapsed = timeit.default_timer() - start_time
    #print('job time (seconds): ', elapsed)
    return pdf

def parallelize_pdf(pdf, func):
    num_cores = 7  #leave one free to not freeze machine
    num_partitions = num_cores #number of partitions to split dataframe\
    
    #print('Population Data before split /n', pdf[:3])
    pdf_split = np.array_split(pdf, num_partitions)
    
    pool = multiprocessing.Pool(num_cores)
    pdf = pd.concat(pool.map(func, pdf_split))
    return pdf

    pool.close()
    pool.join()

#CT Find Road
if __name__=='__main__':
    start_time = timeit.default_timer()
    ct_pop = popPts[popPts.id.str.startswith('09')].copy()
    
    ct_df = parallelize_pdf(ct_pop, add_rdIDs)
    print(ct_df.shape)
    
    elapsed = timeit.default_timer() - start_time
    #print('New Population after pool /n', new_pdf[:3], new_pdf[-3:])
    print('job total time (seconds):', elapsed)
    print('Saving file...')
    ct_df.to_csv('ct_pop_r.csv')

#NJ Find Road
if __name__=='__main__':
    start_time = timeit.default_timer()
    pop = popPts[popPts.id.str.startswith('34')].copy()
    print('Population Size: ', pop.shape)
    ndf = parallelize_pdf(pop, add_rdIDs)
    #Time
    print('After Finnding RoadID',ndf.shape)
    elapsed = timeit.default_timer() - start_time
    #print('New Population after pool /n', new_pdf[:3], new_pdf[-3:])
    print('job total time (seconds):', elapsed)
    print('Saving file...')
    ndf.to_csv('nj_pop_r.csv')

#NY Find Road
if __name__=='__main__':
    start_time = timeit.default_timer()
    pop = popPts[popPts.id.str.startswith('36')].copy()
    print('Population Size: ', pop.shape)
    ndf = parallelize_pdf(pop, add_rdIDs)
    print('After Finnding RoadID',ndf.shape)
    #Time
    elapsed = timeit.default_timer() - start_time
    #print('New Population after pool /n', new_pdf[:3], new_pdf[-3:])
    print('job total time (seconds):', elapsed)
    print('Saving file...')
    ndf.to_csv('ny_pop_r.csv')

ndf.head()

#PA Find Road
if __name__=='__main__':
    start_time = timeit.default_timer()
    new_pdf = parallelize_pdf(papop, add_rdIDs)
    print(new_pdf.shape)
    elapsed = timeit.default_timer() - start_time
    print('New Population after pool /n', new_pdf[:3], new_pdf[-3:])
    print('job total time (seconds):', elapsed)
    #print('Saving file...')
    #new_pdf.to_csv('Out_pop_wrid.csv')
